fix(agent): move markdown skills behind their own subpath
Review found `yaml` was landing in EVERY `@lunora/agent` Worker bundle,
not just those using it: `defineAgent` reaches the module that imports
it. Measured under the Worker build conditions the Cloudflare Vite plugin
uses, minified — 654 bytes with `yaml` external, 34,078 with it bundled.
So ~33 KB of unreachable YAML parser for every app that composes skills
in TypeScript and never opens a markdown file.

`@lunora/agent/skill-markdown` is the seventh subpath this package
already ships, and it puts the cost behind the import that wants it.
Verified: zero `yaml` references in `dist/index.mjs` and the shared
chunks, one in `dist/skill-markdown.mjs`. A lazy `import()` was not an
option — the entry point is synchronous.

Also from review, on the same dependency: a present-but-non-string
`name` reported "found no `name`", sending the author to add a line they
can see is already there, and a frontmatter-only file produced a skill
that merged an empty instruction fragment and did nothing. Both now
throw with the specific reason. The test that "parses list-valued keys"
made the same call as the happy path and asserted a subset of it, so it
could never fail on its own; it now uses a fixture where a line-scanner
would read a nested `name` instead of the document's.

The catalog comment justifying the dependency was wrong on both claims.
`yaml@2.9.0` is `"type": "commonjs"`, and its `node` condition resolves a
CJS build that calls `require("process")`. What makes it workerd-safe is
the RESOLVER, not the package: the plugin's worker conditions match none
of yaml's, so it falls through to the Node-free browser ESM build. An app
adding "node" to its worker `resolve.conditions` would get the other one.

agent 336.
