+++ api-snapshots/agent.api.md
@@ -6,8 +6,8 @@
 Generated by `pnpm run api:update` from the built `dist` types; verified by
 `pnpm run api:check` (CI: Lint / api-surface). Do not edit by hand — a diff
 here is a public-API change and must be reviewed as one. This package is
-Experimental: the snapshot records how its surface moves, and carries NO SemVer
-promise until the package graduates.
+Experimental: the snapshot records how its surface moves, and carries NO
+SemVer promise until the package graduates.
 
 ## `@lunora/agent`
 
@@ -299,10 +299,6 @@ _Tagged `@experimental` — signature not tracked; churn here does not fail the 
 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
 
-### `SkillMarkdownExtras` (type)
-
-_Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
-
 ### `ToolScript` (interface)
 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
@@ -467,10 +463,6 @@ _Tagged `@experimental` — signature not tracked; churn here does not fail the 
 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
 
-### `skillFromMarkdown` (const)
-
-_Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
-
 ### `splitForCompaction` (const)
 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
@@ -628,6 +620,16 @@ _Tagged `@experimental` — signature not tracked; churn here does not fail the 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
 
 ### `fsTool` (const)
+
+_Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
+
+## `@lunora/agent/skill-markdown`
+
+### `SkillMarkdownExtras` (type)
+
+_Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
+
+### `skillFromMarkdown` (const)
 
 _Tagged `@experimental` — signature not tracked; churn here does not fail the gate._
