+++ api-snapshots/ai.api.md
@@ -6,8 +6,8 @@
 Generated by `pnpm run api:update` from the built `dist` types; verified by
 `pnpm run api:check` (CI: Lint / api-surface). Do not edit by hand — a diff
 here is a public-API change and must be reviewed as one. This package is
-Experimental: the snapshot records how its surface moves, and carries NO SemVer
-promise until the package graduates.
+Experimental: the snapshot records how its surface moves, and carries NO
+SemVer promise until the package graduates.
 
 ## `@lunora/ai`
