+++ packages/agent/__tests__/skill.test.ts
@@ -3,18 +3,16 @@ import { describe, expect, it } from "vitest";
 import { runAgentLoop } from "../src/agent-loop";
 import { defineAgent, defineAgentTool } from "../src/define-agent";
 import { DEFAULT_AGENT_FUNCTION_PATHS } from "../src/paths";
-import { defineSkill, isSkillDefinition, skillFromMarkdown } from "../src/skill";
+import { defineSkill, isSkillDefinition } from "../src/skill";
 import type { AgentDefinition, AgentFunctionReference, AgentGenerate, AgentGenerateResult, AgentInstructionsContext, AgentRunFunction } from "../src/types";
 
 const NAME_PATTERN = /identifier/u;
 const COLLIDES_PATTERN = /collides/u;
 const COLLISION_NAMES_PATTERN = /billing.*lookup|lookup.*billing/u;
 const INVALID_IDENTIFIER_PATTERN = /is not a valid identifier/u;
-const RESERVED_NAME_PATTERN = /reserved/u;
 const DUPLICATE_SKILL_PATTERN = /more than one skill/u;
+const RESERVED_NAME_PATTERN = /reserved/u;
 const DUPLICATE_SKILL_NAME_PATTERN = /billing/u;
-const NO_FRONTMATTER_NAME_PATTERN = /no `name` in the markdown frontmatter/u;
-const INVALID_YAML_PATTERN = /invalid YAML frontmatter/u;
 
 /**
  * Faithful in-memory model of Cloudflare Workflows' `step.do` memoization (a
@@ -422,90 +420,5 @@ describe("skill knowledge retrieval in the loop", () => {
 
         expect(journal.invoked.some((name) => name.startsWith("memory:retrieve"))).toBe(false);
         expect(agent.memorySources).toBeUndefined();
-    });
-});
-
-describe(skillFromMarkdown, () => {
-    const SKILL = [
-        "---",
-        "name: triage",
-        "description: not read here, but it must not break the read",
-        "allowed-tools:",
-        "  - bash",
-        "  - read",
-        "---",
-        "Reproduce the report before proposing a cause.",
-        "",
-        "Cite the failing test.",
-    ].join("\n");
-
-    it("takes the name from frontmatter and the instructions from the body", () => {
-        expect.assertions(3);
-
-        const skill = skillFromMarkdown(SKILL);
-
-        expect(skill.name).toBe("triage");
-        expect(skill.instructions).toBe("Reproduce the report before proposing a cause.\n\nCite the failing test.");
-        expect(skill.isLunoraSkill).toBe(true);
-    });
-
-    /**
-     * The reason this parses YAML rather than scanning for one key: a real skill
-     * file carries list- and nested-valued keys meant for other tooling, and a
-     * scalar-only reader would either choke on them or mis-read the `name` that
-     * follows.
-     */
-    it("parses list-valued keys it does not read, instead of choking on them", () => {
-        expect.assertions(1);
-
-        expect(skillFromMarkdown(SKILL).name).toBe("triage");
-    });
-
-    it("merges the code-side extras the file cannot carry", () => {
-        expect.assertions(2);
-
-        const skill = skillFromMarkdown(SKILL, { knowledge: { source: "rag:docs", topK: 2 } });
-
-        expect(skill.knowledge).toStrictEqual({ source: "rag:docs", topK: 2 });
-        expect(skill.name).toBe("triage");
-    });
-
-    it("rejects markdown with no frontmatter, naming what to add", () => {
-        expect.assertions(1);
-
-        expect(() => skillFromMarkdown("Just instructions, no header.")).toThrow(NO_FRONTMATTER_NAME_PATTERN);
-    });
-
-    it("rejects a frontmatter block with no name", () => {
-        expect.assertions(1);
-
-        expect(() => skillFromMarkdown("---\ndescription: nameless\n---\nBody.")).toThrow(NO_FRONTMATTER_NAME_PATTERN);
-    });
-
-    /**
-     * Malformed YAML must not degrade to "no frontmatter" — that reports a
-     * missing `name` for a file whose real problem is a syntax error, and sends
-     * the author to the wrong line.
-     */
-    it("reports invalid YAML as invalid YAML, not as a missing name", () => {
-        expect.assertions(1);
-
-        expect(() => skillFromMarkdown('---\nname: "unterminated\n  bad: [\n---\nBody.')).toThrow(INVALID_YAML_PATTERN);
-    });
-
-    it("applies the same name rules as the object form", () => {
-        expect.assertions(2);
-
-        expect(() => skillFromMarkdown("---\nname: default\n---\nBody.")).toThrow(RESERVED_NAME_PATTERN);
-        expect(() => skillFromMarkdown("---\nname: 9lives\n---\nBody.")).toThrow(NAME_PATTERN);
-    });
-
-    it("tolerates CRLF line endings and a leading BOM", () => {
-        expect.assertions(2);
-
-        const skill = skillFromMarkdown("\uFEFF---\r\nname: triage\r\n---\r\nBody text.\r\n");
-
-        expect(skill.name).toBe("triage");
-        expect(skill.instructions).toBe("Body text.");
     });
 });
